requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
gunicorn
pytz
sqlparse
drf-yasg
docutils
djangorestframework
pytest-django